Update and rename base.py to updated_base.py
The updated code introduces several optimizations to enhance its efficiency and readability. Key changes include:

Refined Imports and Redundancies: Unused imports were removed, and redundant code was simplified to improve clarity and reduce memory usage.

Improved Exception Handling: Exception handling now targets specific errors, providing more informative messages and reducing the likelihood of masking unexpected issues.

Streamlined Methods: Several methods, including those for argument checking and classifier calculation, were optimized to avoid redundancy and complexity, making the code more maintainable.

Enhanced Model Registration: The model registration process was refined to avoid overwriting existing entries and ensure unique model names are correctly registered.

Simplified Error Messaging: Error messages were made more descriptive to aid in debugging, particularly for size-related exceptions.
